Search cryptocurrency functionality
A search form is being used in the site so as to search for specific pages with the use of keywords. For this, a search form is available at the top of all pages and the keywords entered will be used to search for pages. - The search will result in a list of links to all found pages.