Login passwords verification
In this, we have created the function that gives a message if the user information is incorrect while trying to log in. Moreover, if the login attempt is successful ie. the user or admin is able to sign in the website, a message appears saying the login was successful. Users can also log out of the website if required.